kas container: switch to debian bookworm
This patch uses the debian bookworm base image for the kas container. By
that, we can also use the distro version of python3-kconfiglib. In
addition, pip3 on bookworm adds more guards to protect the system from
incompatible global python dependencies. By that, we need to explicitly
tell pip3 install that the way we install kas is safe.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: replace pylint3 with pylint]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>